Refactor database classes location
