--- app/schemas/com.jpaya.englishisfun.database.EnglishIsFunDatabase/2.json
@@ -1,90 +1,0 @@
-{
-  "formatVersion": 1,
-  "database": {
-    "version": 2,
-    "identityHash": "dfded3101abf825e5ea4d8d58a42ebbc",
-    "entities": [
-      {
-        "tableName": "irregulars",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `base` TEXT NOT NULL, `simple` TEXT NOT NULL, `participle` TEXT NOT NULL, `definitions` TEXT NOT NULL)",
-        "fields": [
-          {
-            "fieldPath": "id",
-            "columnName": "id",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "base",
-            "columnName": "base",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "simple",
-            "columnName": "simple",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "participle",
-            "columnName": "participle",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "definitions",
-            "columnName": "definitions",
-            "affinity": "TEXT",
-            "notNull": true
-          }
-        ],
-        "primaryKey": {
-          "columnNames": [
-            "id"
-          ],
-          "autoGenerate": true
-        },
-        "indices": [],
-        "foreignKeys": []
-      },
-      {
-        "tableName": "abbreviations",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `abbr` TEXT NOT NULL, `desc` TEXT NOT NULL)",
-        "fields": [
-          {
-            "fieldPath": "id",
-            "columnName": "id",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "abbr",
-            "columnName": "abbr",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "desc",
-            "columnName": "desc",
-            "affinity": "TEXT",
-            "notNull": true
-          }
-        ],
-        "primaryKey": {
-          "columnNames": [
-            "id"
-          ],
-          "autoGenerate": true
-        },
-        "indices": [],
-        "foreignKeys": []
-      }
-    ],
-    "views": [],
-    "setupQueries": [
-      "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, 'dfded3101abf825e5ea4d8d58a42ebbc')"
-    ]
-  }
-}
