+++ app/src/test/kotlin/com/jpaya/englishisfun/database/di/DatabaseModuleTest.kt
@@ -18,7 +18,8 @@ package com.jpaya.englishisfun.database.di
 
 import com.jpaya.englishisfun.abbreviations.data.db.AbbreviationsDao
 import com.jpaya.englishisfun.conditionals.data.db.ConditionalsDao
-import com.jpaya.englishisfun.database.EnglishIsFunDatabase
+import com.jpaya.englishisfun.data.database.EnglishIsFunDatabase
+import com.jpaya.englishisfun.data.database.di.DatabaseModule
 import com.jpaya.englishisfun.idioms.data.db.IdiomsDao
 import com.jpaya.englishisfun.irregulars.data.db.IrregularsDao
 import com.jpaya.englishisfun.phrasals.data.db.PhrasalsDao
