+++ app/src/test/kotlin/com/jpaya/englishisfun/idioms/data/db/DatabaseDataSourceTest.kt
@@ -17,7 +17,7 @@
 package com.jpaya.englishisfun.idioms.data.db
 
 import androidx.room.Room
-import com.jpaya.englishisfun.database.EnglishIsFunDatabase
+import com.jpaya.englishisfun.data.database.EnglishIsFunDatabase
 import com.jpaya.englishisfun.idioms.domain.Idiom
 import com.jpaya.libraries.testutils.robolectric.TestRobolectric
 import kotlinx.coroutines.runBlocking
